Add pg_log_queries.match_all GUC
By default log_user, log_db and log_addr are checked in this order and
the first matching value will write to the log if log_query is not set
or log_query match. If you want that all of them, when defined, must
match to write to log, activate the match_all directive.  In this case
it will log statements only when defined filters for log_user, log_db,
log_addr and log_query all match.